packages/schema-guesser/pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "schema-guesser"
version = "0.1.0"
description = "Infer JSON Schema Draft 7 from example JSON payloads."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "CLI Tools Contributors" }]
dependencies = [
  "click>=8.1.7",
  "jsonschema>=4.22.0",
  "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[project.scripts]
schema-guesser = "schema_guesser.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/schema_guesser"]
